refactor(renderers): deduplicate renderer code and centralize constants
Extract shared TextAttr, CardAttr, and LanguageDataWithAccum interfaces to
scripts/renderers/types.ts, move theme color magic strings to COLOR_SUBTLE /
COLOR_LIGHT / COLOR_DARK constants in utils/constants.ts, and extract the two
calcPercentages functions into scripts/renderers/calcPercentages.ts.

Closes #39
